adding changes in pr as recomended
